refactor: photo domain 재분리
