+++ frontend/src/components/specific/space/spaceFooter/SpaceFooter.styles.ts
@@ -27,11 +27,3 @@ export const TopButtonContainer = styled.div<{ $isVisible: boolean }>`
     transform: translateX(calc(-50% + 140px));
   }
 `;
-
-export const DownloadButtonContainer = styled.div`
-  position: fixed;
-  bottom: 20px;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: ${({ theme }) => theme.zIndex.floatingActionButton};
-`;
